refactor: 자동차이름 클래스 생성하여 이름 검증 로직 분리
